requirements.txt
six ~= 1.16.0
bagit ~= 1.8.1
exif ~= 1.3.4
boto3 ~= 1.24.35
python-dotenv ~= 0.20.0